pyproject.toml
[project]
name = "pydantic-ai-test"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.116.1",
    "logfire>=4.3.6",
    "pydantic>=2.11.7",
    "pydantic-ai>=0.8.1",
    "python-dotenv>=1.1.1",
]
